Changes to how pizza data is accessed

+++ IS2209CA1ProjectT1/src/main/java/ie/team1/is2209ca1projectt1/dao/PizzaDao.java
@@ -46,7 +46,7 @@ public class PizzaDao {
 
                 int id = rs.getInt("ID");
                 String name = rs.getString("NAME");
-                double price = rs.getInt("PRICE");
+                double price = rs.getDouble("PRICE"); // 29/04/22 changed getInt to getDouble
 
                 Pizza pizza = new Pizza(id, name, price);
                 pizzas.add(pizza);
